Refactor code for better consistency and readability:
- Simplified type imports by removing redundant `std` prefixes.
- Replaced manual error propagation with the `?` operator for cleaner error handling.
- Transitioned from `std::thread::sleep` to `tokio::time::sleep` for consistency with async context.
- Removed redundant `as` casting in favor of direct type literals (e.g., `0f64`) in comparisons.
- Cleaned up unused `TypeId` qualifiers by directly referencing type constants.
- Optimized comparison logic and redundant `unwrap` calls with error handling improvements.
- Enhanced method logic by removing unnecessary return statements and redundant unwraps.

+++ src/buffer/lru_k_replacer.rs
@@ -177,13 +177,13 @@ impl LRUKReplacer {
 
         let entry = store.entry(frame_id).or_insert_with(|| FrameEntry {
             access_count: 0,
-            access_history: VecDeque::with_capacity(self.k as usize),
+            access_history: VecDeque::with_capacity(self.k),
             is_evictable: false,
         });
 
         entry.access_count += 1;
         entry.access_history.push_back(now);
-        if entry.access_history.len() > self.k as usize {
+        if entry.access_history.len() > self.k {
             entry.access_history.pop_front();
         }
